+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 ### Added
+- Added `AnyVec::at` - ergonomic version of `get`.
 - `AnyVecRef` now cloneable.
 - `ElementRef` now cloneable.
 - non-consuming iterators now cloneable.